Revert "Merge remote-tracking branch 'upstream/master' into MORE-secret-doors"
This reverts commit 1660ed293df86c7345640307042b65bb1e6aae53, reversing
changes made to e96187d522decf616a6c3a4189e14ee3d52676af.

+++ Content.Shared/Medical/SuitSensor/SharedSuitSensor.cs
@@ -24,7 +24,7 @@ public sealed class SuitSensorStatus
     public bool IsAlive;
     public int? TotalDamage;
     public int? TotalDamageThreshold;
-    public float? DamagePercentage => TotalDamageThreshold == null || TotalDamage == null ? null : TotalDamage / (float) TotalDamageThreshold;
+    public float? DamagePercentage => TotalDamageThreshold == null || TotalDamage == null ? null : TotalDamage / TotalDamageThreshold;
     public NetCoordinates? Coordinates;
 }
